requirements.txt
PySide6
qdarkstyle
python-chess
